This revision features different diffuse texture types for models on 3D Modelizer. Now, if you click at the Debug Diffuse Texture option, it will replace your texture with the debug one, rather than adding an invisible texture in the model.
There are many other changes in terms of coding. First of all, we'll soon remove the HVA dependence from the model and replace with a more advanced animation module called HierarchyAnimation.pas. I'll also remove the dependence of voxels from Models/LODs/Meshes, so we'll be able to use the rendering code in other programs that may not necessarily involve voxels. 

I've also added Assign commands for several classes, as soon as I've noticed that the Assign from the meshes does not copy its plugins. However, I'm still in doubt if it is worth to make it copy these things, because the mesh is supposed to be independent from its plugins. That doesn't happen at the moment and I'm still studying a way to make it happen.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@447 51f8c993-4936-0410-a6d7-a9c99ad3dd25