This revision copies mesh plugins on assign. There is no difference for the end user, when compared with the previous one.
I've decided to not make plugins independent from the mesh. I couldn't figure out a way to do it. I guess that it might be better to eventually move the voxel support into classes that extends Model, LOD and Mesh or at least one of them. And I don't know if I'll extend the actor to do that either.

Anyway, this revision is considered to be unstable. The last stable revision is still 441. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@448 51f8c993-4936-0410-a6d7-a9c99ad3dd25