This revision improves how VXLSE III deals with OpenCL support.
We've added an OpenCL options under Preferences -> 3D Options. The program also tests the OpenCL support as soon as it runs. I hope nobody gets an access violation every time that the program runs because of that. Although I've added some protection to prevent this kind of issue, the protection was never tested.

Regarding tests, OpenCL seems to be working perfectly on nVidia machines (tested on two so far), but my AMD ATI machine has lost its OpenCL ability, so I'll test this protection from the OpenCL test as soon as I commit this revision.

Although this version is more stable than the previous revision, it is still considered unstable due to issues with the 3D preview window (specially when it closes under some circumstances). The latest stable revision is still 441.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@449 51f8c993-4936-0410-a6d7-a9c99ad3dd25