Models are now independent from voxel models. Our engine can be now used on non-voxel programs to render non-voxel related models.
This revision ensures this difference by removing several dependencies of voxels and HVas from TModel, TLD and TMesh. The new THierarchyAnimation is now used instead of HVA there. It has TModelVxt and TMeshVxt for Voxel Model support. There is also a VOXEL_SUPPORT mode that can be disabled at Global_Conditionals.inc for full voxel deactivation.

The voxel independence has also interefered on the RenderEnvironment as the program no longer counts polygons and voxels on every frame that it renders and it may not necessarily display this data at the bottom of the screen.

For the end user, what changes is:

- The main renderer only displays the amount of voxels in the model (it no longer displays number of faces).
- 3D Modelizer also displays the amount of vertices. 3D Preview displays both polycount and voxel count.
- The models may be rendered faster (nearly un-noticeable by humans).
- 3D Preview window is now synchronized again with what is being edited.
- Fixed a bug at the Voxel Header window that could cause an error message when editing the transformation with an invalid number.


This revision is still considered unstable, although it is not that much unstable anymore. The latest stable revision is 441.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@450 51f8c993-4936-0410-a6d7-a9c99ad3dd25