This revision fix a couple of animation issues introduced by the last patch. The THierarchyAnimation object processed the transformation matrix in a inverted way, providing very weird results when the matrix wasn't identity. Also, frames were not being animated... and there are more bugs.
This revision doesn't fix all of them, because I'm tired right now. But it improves it by animating frames (the frame number displayed on 3D Preview and 3D Modelizer are not reliable yet, neither the Play, Pause and Stop icons). The position of the meshes are wrong as well, as far as I've tested the Mammoth Mark II voxel model. Finally, there is a bug that causes an invalid pointer operation if you try to close the program with both 3D Preview and 3D Modelizer opened. I have no idea why it gives me an invalid pointer operation when ending the command list from the controller.

For this (and few other reasons), this revision is not stable. The latest stable revision is 441.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@451 51f8c993-4936-0410-a6d7-a9c99ad3dd25