This revision fixes an invalid pointer operator bug with the controller, when it removes an item. This was the most annoying bug that prevented the last 10 or more revisions from being considered stable.
This revision is still considered unstable, but that happens due to the issues with the Frame combo box on 3D Preview window and Modelizer. It doesn't change the frame of the animation at the moment. I'll fix it in the next hours or days.

The latest stable revision is still 441.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@453 51f8c993-4936-0410-a6d7-a9c99ad3dd25