This revision fixes the triangle collision algorithm used in both Texture Atlas Extraction Origami and Origami Parametric. By fixing, it means it damn works correctly, finally. Both CPU and OpenCL implementations were fixed in this revision.
If you have a previous revision of VXLSE III with origami.cl, you'll also need to update this file. The quickest way to do it is to simply delete the file and let Voxel Section Editor III's auto installer re-download it once you run the program. If you use an old origami.cl, these techniques will generate wrong triangles or restrict it from creating valid triangles.

The Origami texture atlas extractor still has an non-identified bug that prevents it from adding all faces in some cases or some faces just become absurdly small. What's happening there is not clear for me yet.

The parametric origami is a waste of time and it was only useful to force me to fix the triangle collision algorithm. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@462 51f8c993-4936-0410-a6d7-a9c99ad3dd25