Additional bug fixes on Origami Texture Atlas Extractor (triangle collisions and curvature measures).
In the last revision I've mentioned that the triangle collision detection algorithm used on the Origami Texture Atlas Extractor was correct. I was wrong. This time, I won't say that this triangle collision algorithm is correct, but what I can say is that there is a strong theory behind it already and I have not detected any problems with it... yet.

There are other fixes such as numeric issues on VertexTransformationUtils.pas and ordering issues on NeighborDetector.pas. They are related to the way it orders the vertexes around a vertex according to its angles. The results generated by this operation are better now and that has improved a lot the results from the Origami Parametric Texture Atlas Extractor method (which is still worse than Origami). 

Also, I've removed a SetVector function from BasicFunctions.pas. This function already exists at Math3D. This is why some other files were modified.

In order to use the latest origami.cl to run the Origami texture atlas extractor method in GPU, make sure you delete your existing one and let the program download the new one. The Origami method generates interesting texture maps as long as your model does not have any ceil critical point. It is a very rare thing for detailed models. And that's our next target for the Origami method.  

git-svn-id: svn://svn.ppmsite.com/voxel_tools@463 51f8c993-4936-0410-a6d7-a9c99ad3dd25