This revision fixes two problems reported by Lin Kuei Ominae: http://www.ppmsite.com/forum/viewtopic.php?p=523908#523908
I've noticed that there was indeed a mistake at the ordering of the views (Left and Right, Top and Bottom) regarding the axis system used by the program. So, they should look correctly now.

LKO has also pointed out one of the obscure sections of the program that I did not ever touch on it and it suffered the pain of the different axis system of the program. The dimension of the New Sections should now match what has been specified by the user.

Finally, the Voxel_Engine.pas has received an improved identation at the code from HBD.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@489 51f8c993-4936-0410-a6d7-a9c99ad3dd25