This revision fixes some issues at the Auto Normals dialog. It mostly improves hints, fixes spelling, etc. Despite the complaints from G-E at the link below, I've preferred to keep the Tangent Plane Auto Normals as the default one.
http://www.ppmsite.com/forum/viewtopic.php?t=39294

The truth is that all these auto normals methods are not adequate because all of them have a local approach instead of global.

Also, we have fixed range (that is determined by the user). So, all of them fail to find the appropriate sampling (normalization range, in the language of the program) required to find the normal vector, when every voxel is actually a unique surface (and does not have multiples 'correct' normal vectors). 

What happens with big ranges is that their surface may have discontinuity, which generate noises in the result and terrible results. Small ranges may also contribute with bad results, since we loose information that could give us the correct normal direction, but the distortion is not as bad as the regions with discontinuity. This is why smaller ranges usually get better results than very big ranges.

Considering that we are dealing with surfaces that are part of objects with different curvatures, volumes, etc... and the range must vary according to these properties, a fixed range on its own is not adequate either.

Anyway, I do not have plans to build a better auto normals algorithm at the moment. So, the recommendation is to use Influence Normals for rounded objects or Tangent Planes for sharper ones.


Regarding the Actor.pas, I've just undone something that I've messed up in few revisions ago. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@490 51f8c993-4936-0410-a6d7-a9c99ad3dd25