OS: HVA Builder allows the selection of Lepton Size and Bullet Size, at the Preferences.
By default, Lepton Size is 43/256 (Mig Eater and LKO), although you can easily change to 45/256 (Stucuk and GE). I personally prefer Mig's value. It works lightly better for IFV and few other units. But it is hard to compare the results, since the difference between both values is minimal. 

Regarding the bullet multiplier size, the default value is 5. It means that each edge is 5 times bigger than the edge of one voxel in the scene. OS: Voxel Viewer wasn't really changed this time and it still uses Mig's value for lepton size and bullet multiplier size of 5.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@575 51f8c993-4936-0410-a6d7-a9c99ad3dd25